login/signup routes fixed and added signup page
