+++ controllers/api/userRoutes.js
@@ -2,6 +2,7 @@ const router = require('express').Router();
 const { User } = require('../../models');
 
 router.post('/', async (req, res) => {
+  console.log("Hello");
   try {
     const userData = await User.create(req.body);
