+++ views/layouts/main.handlebars
@@ -14,7 +14,7 @@
     <header class="display-flex justify-space-between align-center p-2">
       <h1>
         <span role="img" aria-label="money">💸</span>
-        <a href="/">Fund My Project!</a>
+        <a href="/">Fitness Pal</a>
       </h1>
       <nav>
         <!-- Conditionally render login or logout links -->
